Timing safe string compare function
The implementation for `CompareStrings` function initially used strcmp,
which is not timing safe, i.e. it breaks at the first non-equal
character which allows information to be extracted about the prefix that
is matching, which is know as the timing attack. To avoid that, we use
the modified `CompareStrings`, that will iterate over all the
character until the very end of the supplied password only. So, in this
way all operations will happen exactly the length of supplied string s1
times.